# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almacen API",
        "description": "API para administrar items, categorias, proveedores, ubicaciones y movimientos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movements": {
            "get": {
                "tags": [
                    "Movements"
                ],
                "summary": "List movements",
                "description": "Display a listing of the resource.",
                "operationId": "12b7c3a98a31664a27f614325987684e",
                "responses": {
                    "200": {
                        "description": "Movements retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Movements"
                ],
                "summary": "Create a movement",
                "description": "Store a newly created resource in storage.",
                "operationId": "9d84e9d5ed2d562b43ed79402fe3eb5e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_id",
                                    "user_id",
                                    "type",
                                    "quantity"
                                ],
                                "properties": {
                                    "item_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "entrada",
                                            "salida"
                                        ]
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "note": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movement created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/movements/{movement}": {
            "get": {
                "tags": [
                    "Movements"
                ],
                "summary": "Show a movement",
                "description": "Display the specified resource.",
                "operationId": "d61c22c71da0c4b20c0de4888253f18a",
                "parameters": [
                    {
                        "name": "movement",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movement retrieved successfully"
                    },
                    "404": {
                        "description": "Movement not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movements"
                ],
                "summary": "Update a movement",
                "description": "Update the specified resource in storage.",
                "operationId": "72e76ef77740ea46834c55790e069a13",
                "parameters": [
                    {
                        "name": "movement",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "item_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "entrada",
                                            "salida"
                                        ]
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "note": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movement updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movements"
                ],
                "summary": "Delete a movement",
                "description": "Remove the specified resource from storage.",
                "operationId": "f34c121777b9bb352b2a298f720ad5ae",
                "parameters": [
                    {
                        "name": "movement",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movement deleted successfully"
                    }
                }
            }
        },
...
    },
...
}